+++ src/connectors/impls/gcs.rs
@@ -12,4 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod api_client;
+mod chunked_buffer;
 pub(crate) mod writer;
